opt step 5: switch to bfloat16

+++ train_optimized.py
@@ -20,7 +20,7 @@ def main():
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
     model = GPTModel(gpt_config)
-    model.to(device)
+    model.to(device).to(torch.bfloat16)  # ! optimized step 5: use bfloat16 precision
 
     optimizer = torch.optim.AdamW(
         model.parameters(),
